feat: Button support lineHeight token
