+++ components/button/style/index.ts
@@ -1,162 +1,14 @@
-import type { CSSProperties } from 'react';
 import type { CSSInterpolation, CSSObject } from '@ant-design/cssinjs';
 import { unit } from '@ant-design/cssinjs';
 
 import { genFocusStyle } from '../../style';
-import type { FullToken, GenerateStyle, GetDefaultToken } from '../../theme/internal';
+import type { GenerateStyle } from '../../theme/internal';
 import { genStyleHooks, mergeToken } from '../../theme/internal';
-import type { GenStyleFn } from '../../theme/util/genComponentStyleHook';
 import genGroupStyle from './group';
+import type { ButtonToken, ComponentToken } from './token';
+import { formatComponentToken, prepareComponentToken, prepareToken } from './token';
 
-/** Component only token. Which will handle additional calculation of alias token */
-export interface ComponentToken {
-  /**
-   * @desc 文字字重
-   * @descEN Font weight of text
-   */
-  fontWeight: CSSProperties['fontWeight'];
-  /**
-   * @desc 默认按钮阴影
-   * @descEN Shadow of default button
-   */
-  defaultShadow: string;
-  /**
-   * @desc 主要按钮阴影
-   * @descEN Shadow of primary button
-   */
-  primaryShadow: string;
-  /**
-   * @desc 危险按钮阴影
-   * @descEN Shadow of danger button
-   */
-  dangerShadow: string;
-  /**
-   * @desc 主要按钮文本颜色
-   * @descEN Text color of primary button
-   */
-  primaryColor: string;
-  /**
-   * @desc 默认按钮文本颜色
-   * @descEN Text color of default button
-   */
-  defaultColor: string;
-  /**
-   * @desc 默认按钮背景色
-   * @descEN Background color of default button
-   */
-  defaultBg: string;
-  /**
-   * @desc 默认按钮边框颜色
-   * @descEN Border color of default button
-   */
-  defaultBorderColor: string;
-  /**
-   * @desc 危险按钮文本颜色
-   * @descEN Text color of danger button
-   */
-  dangerColor: string;
-  /**
-   * @desc 禁用状态边框颜色
-   * @descEN Border color of disabled button
-   */
-  borderColorDisabled: string;
-  /**
-   * @desc 默认幽灵按钮文本颜色
-   * @descEN Text color of default ghost button
-   */
-  defaultGhostColor: string;
-  /**
-   * @desc 幽灵按钮背景色
-   * @descEN Background color of ghost button
-   */
-  ghostBg: string;
-  /**
-   * @desc 默认幽灵按钮边框颜色
-   * @descEN Border color of default ghost button
-   */
-  defaultGhostBorderColor: string;
-  /**
-   * @desc 按钮横向内间距
-   * @descEN Horizontal padding of button
-   */
-  paddingInline: CSSProperties['paddingInline'];
-  /**
-   * @desc 大号按钮横向内间距
-   * @descEN Horizontal padding of large button
-   */
-  paddingInlineLG: CSSProperties['paddingInline'];
-  /**
-   * @desc 小号按钮横向内间距
-   * @descEN Horizontal padding of small button
-   */
-  paddingInlineSM: CSSProperties['paddingInline'];
-  /**
-   * @desc 按钮横向内间距
-   * @descEN Horizontal padding of button
-   */
-  paddingBlock: CSSProperties['paddingInline'];
-  /**
-   * @desc 大号按钮横向内间距
-   * @descEN Horizontal padding of large button
-   */
-  paddingBlockLG: CSSProperties['paddingInline'];
-  /**
-   * @desc 小号按钮横向内间距
-   * @descEN Horizontal padding of small button
-   */
-  paddingBlockSM: CSSProperties['paddingInline'];
-  /**
-   * @desc 只有图标的按钮图标尺寸
-   * @descEN Icon size of button which only contains icon
-   */
-  onlyIconSize: number;
-  /**
-   * @desc 大号只有图标的按钮图标尺寸
-   * @descEN Icon size of large button which only contains icon
-   */
-  onlyIconSizeLG: number;
-  /**
-   * @desc 小号只有图标的按钮图标尺寸
-   * @descEN Icon size of small button which only contains icon
-   */
-  onlyIconSizeSM: number;
-  /**
-   * @desc 按钮组边框颜色
-   * @descEN Border color of button group
-   */
-  groupBorderColor: string;
-  /**
-   * @desc 链接按钮悬浮态背景色
-   * @descEN Background color of link button when hover
-   */
-  linkHoverBg: string;
-  /**
-   * @desc 文本按钮悬浮态背景色
-   * @descEN Background color of text button when hover
-   */
-  textHoverBg: string;
-  /**
-   * @desc 按钮内容字体大小
-   * @descEN Font size of button content
-   */
-  contentFontSize: number;
-  /**
-   * @desc 大号按钮内容字体大小
-   * @descEN Font size of large button content
-   */
-  contentFontSizeLG: number;
-  /**
-   * @desc 小号按钮内容字体大小
-   * @descEN Font size of small button content
-   */
-  contentFontSizeSM: number;
-}
-
-export interface ButtonToken extends FullToken<'Button'> {
-  buttonPaddingHorizontal: CSSProperties['paddingInline'];
-  buttonPaddingVertical: CSSProperties['paddingBlock'];
-  buttonIconOnlyFontSize: number;
-}
+export type { ComponentToken };
 
 // ============================== Shared ==============================
 const genSharedButtonStyle: GenerateStyle<ButtonToken, CSSObject> = (token): CSSObject => {
@@ -177,7 +29,6 @@ const genSharedButtonStyle: GenerateStyle<ButtonToken, CSSObject> = (token): CSS
       transition: `all ${token.motionDurationMid} ${token.motionEaseInOut}`,
       userSelect: 'none',
       touchAction: 'manipulation',
-      lineHeight: token.lineHeight,
       color: token.colorText,
 
       '&:disabled > *': {
@@ -547,6 +398,7 @@ const genSizeButtonStyle = (token: ButtonToken, sizePrefixCls: string = ''): CSS
     componentCls,
     controlHeight,
     fontSize,
+    lineHeight,
     borderRadius,
     buttonPaddingHorizontal,
     iconCls,
@@ -560,6 +412,7 @@ const genSizeButtonStyle = (token: ButtonToken, sizePrefixCls: string = ''): CSS
     {
       [`${componentCls}${sizePrefixCls}`]: {
         fontSize,
+        lineHeight,
         height: controlHeight,
         padding: `${unit(buttonPaddingVertical!)} ${unit(buttonPaddingHorizontal!)}`,
         borderRadius,
@@ -602,6 +455,7 @@ const genSizeBaseButtonStyle: GenerateStyle<ButtonToken> = (token) =>
   genSizeButtonStyle(
     mergeToken<ButtonToken>(token, {
       fontSize: token.contentFontSize,
+      lineHeight: token.contentLineHeight,
     }),
   );
 
@@ -609,6 +463,7 @@ const genSizeSmallButtonStyle: GenerateStyle<ButtonToken> = (token) => {
   const smallToken = mergeToken<ButtonToken>(token, {
     controlHeight: token.controlHeightSM,
     fontSize: token.contentFontSizeSM,
+    lineHeight: token.contentLineHeightSM,
     padding: token.paddingXS,
     buttonPaddingHorizontal: token.paddingInlineSM,
     buttonPaddingVertical: token.paddingBlockSM,
@@ -623,6 +478,7 @@ const genSizeLargeButtonStyle: GenerateStyle<ButtonToken> = (token) => {
   const largeToken = mergeToken<ButtonToken>(token, {
     controlHeight: token.controlHeightLG,
     fontSize: token.contentFontSizeLG,
+    lineHeight: token.contentLineHeightLG,
     buttonPaddingHorizontal: token.paddingInlineLG,
     buttonPaddingVertical: token.paddingBlockLG,
     borderRadius: token.borderRadiusLG,
@@ -644,52 +500,6 @@ const genBlockButtonStyle: GenerateStyle<ButtonToken> = (token) => {
 };
 
 // ============================== Export ==============================
-export const prepareToken: (token: Parameters<GenStyleFn<'Button'>>[0]) => ButtonToken = (
-  token,
-) => {
-  const { paddingInline, onlyIconSize, paddingBlock } = token;
-
-  const buttonToken = mergeToken<ButtonToken>(token, {
-    buttonPaddingHorizontal: paddingInline,
-    buttonPaddingVertical: paddingBlock,
-    buttonIconOnlyFontSize: onlyIconSize,
-  });
-
-  return buttonToken;
-};
-
-export const prepareComponentToken: GetDefaultToken<'Button'> = (token) => ({
-  fontWeight: 400,
-  defaultShadow: `0 ${token.controlOutlineWidth}px 0 ${token.controlTmpOutline}`,
-  primaryShadow: `0 ${token.controlOutlineWidth}px 0 ${token.controlOutline}`,
-  dangerShadow: `0 ${token.controlOutlineWidth}px 0 ${token.colorErrorOutline}`,
-  primaryColor: token.colorTextLightSolid,
-  dangerColor: token.colorTextLightSolid,
-  borderColorDisabled: token.colorBorder,
-  defaultGhostColor: token.colorBgContainer,
-  ghostBg: 'transparent',
-  defaultGhostBorderColor: token.colorBgContainer,
-  paddingInline: token.paddingContentHorizontal - token.lineWidth,
-  paddingInlineLG: token.paddingContentHorizontal - token.lineWidth,
-  paddingInlineSM: 8 - token.lineWidth,
-  paddingBlock: 0,
-  paddingBlockSM: 0,
-  paddingBlockLG: 0,
-  onlyIconSize: token.fontSizeLG,
-  onlyIconSizeSM: token.fontSizeLG - 2,
-  onlyIconSizeLG: token.fontSizeLG + 2,
-  groupBorderColor: token.colorPrimaryHover,
-  linkHoverBg: 'transparent',
-  textHoverBg: token.colorBgTextHover,
-  defaultColor: token.colorText,
-  defaultBg: token.colorBgContainer,
-  defaultBorderColor: token.colorBorder,
-  defaultBorderColorDisabled: token.colorBorder,
-  contentFontSize: token.fontSize,
-  contentFontSizeSM: token.fontSize,
-  contentFontSizeLG: token.fontSizeLG,
-});
-
 export default genStyleHooks(
   'Button',
   (token) => {
@@ -718,20 +528,10 @@ export default genStyleHooks(
   {
     unitless: {
       fontWeight: true,
+      contentLineHeight: true,
+      contentLineHeightSM: true,
+      contentLineHeightLG: true,
     },
-    format: (token) => ({
-      paddingBlock: Math.max(
-        (token.controlHeight - token.contentFontSize * token.lineHeight) / 2 - token.lineWidth,
-        0,
-      ),
-      paddingBlockSM: Math.max(
-        (token.controlHeightSM - token.contentFontSizeSM * token.lineHeight) / 2 - token.lineWidth,
-        0,
-      ),
-      paddingBlockLG: Math.max(
-        (token.controlHeightLG - token.contentFontSizeLG * token.lineHeight) / 2 - token.lineWidth,
-        0,
-      ),
-    }),
+    format: formatComponentToken,
   },
 );
