+++ components/theme/themes/shared/genFontSizes.ts
@@ -1,4 +1,8 @@
 // https://zhuanlan.zhihu.com/p/32746810
+export function getLineHeight(fontSize: number) {
+  return (fontSize + 8) / fontSize;
+}
+
 export default function getFontSizes(base: number) {
   const fontSizes = new Array(10).fill(null).map((_, index) => {
     const i = index - 1;
@@ -11,12 +15,8 @@ export default function getFontSizes(base: number) {
 
   fontSizes[1] = base;
 
-  return fontSizes.map((size) => {
-    const height = size + 8;
-
-    return {
-      size,
-      lineHeight: height / size,
-    };
-  });
+  return fontSizes.map((size) => ({
+    size,
+    lineHeight: getLineHeight(size),
+  }));
 }
